feat: added mui and modular architecture
