+++ react-basics/solutions/todoApp/src/components/TodoForm/TodoForm.types.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+import { Todo } from "../../types/todo";
+import { createTodoSchema } from "../../schemas/createTodoSchema";
+
+export type TodoFormProps = {
+  todos: Todo[];
+  setTodos: React.Dispatch<React.SetStateAction<Todo[]>>;
+};
+
+export type CreateTodoSchema = z.infer<typeof createTodoSchema>;
