+++ react-basics/solutions/todoApp/src/components/TodoForm/index.ts
@@ -1,0 +1,3 @@
+import { TodoForm } from "./TodoForm";
+
+export { TodoForm };
