+++ react-basics/solutions/todoApp/src/components/TodoItem/TodoItem.tsx
@@ -1,6 +1,6 @@
-import { Todo } from "../types/todo";
+import { Todo } from "../../types/todo";
 
-export function TodoView({ todo }: { todo: Todo }) {
+export function TodoItem({ todo }: { todo: Todo }) {
   return (
     <div>
       <p className={todo.completed ? "line-through text-gray-600" : ""}>
