+++ react-basics/solutions/todoApp/src/components/TodoItem/index.ts
@@ -1,0 +1,3 @@
+import { TodoItem } from "./TodoItem";
+
+export { TodoItem };
