+++ react-basics/solutions/todoApp/src/components/TodoList/TodoList.types.ts
@@ -1,0 +1,6 @@
+import { Todo } from "../../types/todo";
+
+export type TodoListProps = {
+  todos: Todo[];
+  setTodos: React.Dispatch<React.SetStateAction<Todo[]>>;
+};
