+++ react-basics/solutions/todoApp/src/components/TodoList/index.ts
@@ -1,0 +1,3 @@
+import { TodoList } from "./TodoList";
+
+export { TodoList };
