+++ react-basics/solutions/todoApp/src/components/TodoMessage/index.ts
@@ -1,0 +1,3 @@
+import { TodoMessage } from "./TodoMessage";
+
+export { TodoMessage };
