+++ react-basics/solutions/todoApp/src/schemas/createTodoSchema.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
 
-export const TodoSchema = z.object({
+export const createTodoSchema = z.object({
   name: z.string(),
 });
